back_processor/mailing_system/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg[binary]
python-dotenv
faiss-cpu
sentence-transformers
google-genai
pypdf



python-docx
pdfplumber
pydantic
python-jose[cryptography]
pydantic-settings
python-json-logger
